Process Bug Fixing
This fix remove the loading in the project box and directly opens the
log view. The process system has been changed to ensure all messages
from the child process are catched and now are displayed in real time to
the log view. This increase the functionality considerably because in
development time Meteor shows compilation errors directly in the console
and they were not updated in the app log view after the project was
launched.